wcx: improve the convenience script
This little script can now accept one or more file names or paths as an
argument. Then it simply executes the wc command with the appropriate
option and arguments. Again this is all to minimize the number of keystrokes.

Also, the user can now get the char and word count from wc by calling either wcm
or wcw respectively. New symbolic links will have to be created to the wcx shell
script to use this new functionality.